% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-main.R
\docType{package}
\name{BMSC-package}
\alias{BMSC-package}
\title{BMSC: Bayesian Multilevel Single Case models using 'Stan'}
\description{
The \strong{BMSC} package provides an interface to fit Bayesian Multilevel Single Case models.
These models compare the performance of a single patient against a control group, combining
the flexibility of multilevel models and the potentiality of Bayesian Statistics.
}
\details{
The package is now limited to gaussian data only, but we will further expand it to cover
binomial and ordinal (Likert scales) data.

By means of BMSC the effects of the control group and the effects of the deviance between the
patient and the group will be esimated.

The model to estimate the controls parameters is:

\deqn{y~N(\beta X + b Z, \sigma^2)}

where \eqn{y} is the controls' dependent variable, \eqn{X} the contrast matrix for Population-level (or Fixed)
Effects, and \eqn{\beta} are the unknown coefficients to be estimate. \eqn{Z} is the contrast matrix for the
Varying (or Random, or Group-level) effects, and \eqn{b} are the unknown estimates for the varying effects.
\eqn{\sigma^2} is the variance.

In order to estimate the coefficients of the patient, the formula is the following:

\deqn{y_{pt}~\mathcal{N}(\phi X_{pt}, \sigma_{pt}^2)}

where \eqn{\phi = \beta + \delta}.
}
\section{Details}{

The main function of \strong{BMSC} is \code{\link{BMSC}}, which uses formula syntax to
specify your model.
}

