% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-main.R
\docType{package}
\name{BMSC}
\alias{BMSC}
\title{BMSC: Bayesian Multilevel Single Case models using 'Stan'}
\usage{
BMSC(formula, data.ctrl, data.pt, cores = 1, chains = 4, warmup = 2000,
            iter = 4000, seed = NA, ...)
}
\arguments{
\item{formula}{An object of class \code{formula}: a symbolic description of the model to be fitted.}

\item{data.ctrl}{An object of class \code{data.frame} (or one that can be coerced to that class)
containing data of all variables used in the model for the control group.}

\item{data.pt}{An object of class \code{data.frame} (or one that can be coerced to that class)
containing data of all variables used in the model for the patient.}

\item{chains}{Number of Markov chains (defaults to 4).}

\item{warmup}{A positive integer specifying number of warmup (aka burnin) iterations.
This also specifies the number of iterations used for stepsize adaptation,
so warmup samples should not be used for inference. The number of warmup should
not be larger than iter and the default is 2000.}

\item{iter}{Number of total iterations per chain (including warmup; defaults to 4000).}

\item{seed}{The seed for random number generation to make results reproducible.
If NA (the default), Stan will set the seed randomly.}

\item{typeprior}{Set the desired prior distribution for the fixed effects.
\describe{
     \item{normal} a normal distribution with \mu = 0 and \sigma = 10
     \item{cauchy} a cauchy distribution with \mu = 0 and scale \sqrt(2)/2
     \item{student} a Student's T distribution, with \mu = 0, \nu = 3 and \sigma = 10
 }
 The normal distribution is the default.}

\item{...}{further arguments to be passed to \strong{stan} function.}
}
\description{
The \strong{BMSC} package provides an interface to fit Bayesian Multilevel Single Case models.
These models compare the performance of a single patient against a control group, combining
the flexibility of multilevel models and the potentiality of Bayesian Statistics.

\code{bmscfit} fits the Bayesian Multilevel Single Case models.
}
\details{
The package is now limited to gaussian data only, but we will further expand it to cover
binomial and ordinal (Likert scales) data.

By means of BMSC the effects of the control group and the effects of the deviance between the
patient and the group will be esimated.

The model to estimate the controls parameters is:

\deqn{y~\mathcal{N}(\beta X + b Z, \sigma^2)}

where $y$ is the controls' dependent variable, $X$ the contrast matrix for Population-level (or Fixed)
Effects, and $\beta$ are the unknown coefficients to be estimate. $Z$ is the contrast matrix for the
Varying (or Random, or Group-level) effects, and $b$ are the unknown estimates for the varying effects.
$\sigma^2$ is the variance.

In order to estimate the coefficients of the patient, the formula is the following:

\deqn{y_{pt}~\mathcal{N}(\phi X_{pt}, \sigma_{pt}^2)}

where $\phi = \beta + \delta$.
}
\section{Details}{

The main function of \strong{BMSC} is \code{\link{BMSC}}, which uses formula syntax to
specify your model.
}

