% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-loo.R
\name{BMSC_loo_compare}
\alias{BMSC_loo_compare}
\alias{print.loo_compare_BMSC}
\alias{print.waic_compare_BMSC}
\title{bmscstan wrapper for model comparison.}
\usage{
BMSC_loo_compare(x, ...)

\method{print}{loo_compare_BMSC}(x, simplify = TRUE, ...)

\method{print}{waic_compare_BMSC}(x, simplify = TRUE, ...)
}
\arguments{
\item{x}{A list of \code{loo_BMSC} or  \code{waic_BMSC} objects.}

\item{...}{further arguments passed to the function.}

\item{simplify}{For the print method only, should only the essential columns
of the summary matrix be printed? The entire matrix is always returned,
but by default only the most important columns are printed.}
}
\value{
a list with the log likelihood of the single case and the
        control group,
        the MCMC effective sample size divided by the total sample size,
        and the leave-one-out cross-validation.
}
\description{
bmscstan wrapper for model comparison.
}
