% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-main.R, R/BMSC-random_effects.R
\docType{package}
\name{bmscstan}
\alias{bmscstan}
\title{Bayesian Multilevel Single Case models using 'Stan'}
\description{
The \strong{bmscstan} package provides an interface to fit Bayesian
Multilevel Single Case models.
These models compare the performance of a Single Case against a control
group, combining
the flexibility of multilevel models and the potentiality of Bayesian
Statistics.

The \strong{BMSC} function allows the flexibility of multilevel (generalised)
linear models on single case analysis.
}
\details{
The package is now limited to gaussian data only, but we will further expand
it to cover
binomial and ordinal (Likert scales) data.

By means of \strong{bmscstan} the effects of the control group and the
effects of the deviance between the
Single Case and the group will be esimated.

The model to estimate the controls parameters is:

\ifelse{html}{\out{<center><i>y~N(&beta; X + b Z, &sigma;<sup>2</sup>)</i></center>}}{\eqn{y~N(\beta X + b Z, \sigma^2)}}


where \eqn{y} is the controls' dependent variable, \eqn{X} the contrast
matrix for Population-level (or Fixed)
Effects, and \eqn{\beta} are the unknown coefficients to be estimate. \eqn{Z}
is the contrast matrix for the
Varying (or Random, or Group-level) effects, and \eqn{b} are the unknown
estimates for the varying effects.
\eqn{\sigma^2} is the variance.

In order to estimate the coefficients of the Single Case, the formula is the following:

\ifelse{html}{\out{<center><i>y<sub>pt</sub>~N(&phi; X<sub>pt</sub>, &sigma;<sup>2</sup><sub>pt</<sub>)</i></center>}}{\eqn{y_{pt}~N(\phi X_{pt}, \sigma_{pt}^2)}}

where \eqn{\phi = \beta + \delta}.

In particular, it is possible to specify the population-level (a.k.a. mixed
effects) and the group-level (a.k.a. random effects) coefficients.

The specification of the population- and group-level effects can be done
using the well-known \strong{lme4} notation with specific limitations:

* it is no possible to estimate uncorrelated group-level effects
* it is no possible to directly estimate nested effects. You need to use
  a specific trick

However, in the specification of group-level effects

package provides an interface to fit Bayesian
Multilevel Single Case models.
These models compare the performance of a Single Case against a control
group, combining
the flexibility of multilevel models and the potentiality of Bayesian
Statistics.

The package is now limited to gaussian data only, but we will further expand
it to cover
binomial and ordinal (Likert scales) data.

By means of \strong{bmscstan} the effects of the control group and the
effects of the deviance between the
Single Case and the group will be esimated.

The model to estimate the controls parameters is:

\ifelse{html}{\out{<center><i>y~N(&beta; X + b Z, &sigma;<sup>2</sup>)</i></center>}}{\eqn{y~N(\beta X + b Z, \sigma^2)}}


where \eqn{y} is the controls' dependent variable, \eqn{X} the contrast
matrix for Population-level (or Fixed)
Effects, and \eqn{\beta} are the unknown coefficients to be estimate. \eqn{Z}
is the contrast matrix for the
Varying (or Random, or Group-level) effects, and \eqn{b} are the unknown
estimates for the varying effects.
\eqn{\sigma^2} is the variance.

In order to estimate the coefficients of the Single Case, the formula is the following:

\ifelse{html}{\out{<center><i>y<sub>pt</sub>~N(&phi; X<sub>pt</sub>, &sigma;<sup>2</sup><sub>pt</<sub>)</i></center>}}{\eqn{y_{pt}~N(\phi X_{pt}, \sigma_{pt}^2)}}

where \eqn{\phi = \beta + \delta}.
}
\section{Details}{

The main function of \strong{bmscstan} is \code{\link{BMSC}}, which uses formula syntax to
specify your model.


The main function of \strong{bmscstan} is \code{\link{BMSC}}, which uses formula syntax to
specify your model.
}

