% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-main.R
\name{bridge_sampler.BMSC}
\alias{bridge_sampler.BMSC}
\title{Computes log marginal likelihood via bridge sampling.}
\usage{
bridge_sampler.BMSC(object, ...)
}
\arguments{
\item{object}{a BMSC object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
an "psis_loo" "loo" object
}
\description{
Computes log marginal likelihood via bridge sampling.
}
