% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-main.R
\docType{data}
\name{data.ctrl}
\alias{data.ctrl}
\title{Data from a control group of 16 participants}
\format{
A data frame with 4049 rows and 5 variables
\describe{
 \item{RT}{Reaction times, in milliseconds}
 \item{Body.District}{Body district, categorial factor of
       Body Sidedness Task: FOOT or HAND}
 \item{Congruency}{The trail was Congruent or Incongruent?}
 \item{Side}{The trial showed a left or right limb}
 \item{ID}{The participant ID}
}
}
\usage{
data.ctrl
}
\description{
A dataset containing the results from the Body Sidednedd Task
from a control group of 16 participants
}
\keyword{datasets}
