% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-main.R
\docType{data}
\name{data.pt}
\alias{data.pt}
\title{Data from a Single Case with brachial plexious lesion}
\format{
A data frame with 467 rows and 4 variables
\describe{
 \item{RT}{Reaction times, in milliseconds}
 \item{Body.District}{Body district, categorial factor of
       Body Sidedness Task: FOOT or HAND}
 \item{Congruency}{The trail was Congruent or Incongruent?}
 \item{Side}{The trial showed a left or right limb}
}
}
\usage{
data.pt
}
\description{
A dataset containing the results from the Body Sidedness Task
from a single Single Case
}
\keyword{datasets}
