% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-loo.R
\name{loo.BMSC}
\alias{loo.BMSC}
\title{bmscstan wrapper for computing approximate leave-one-out cross-validation
using PSIS-LOO
for the single case and the control group}
\usage{
\method{loo}{BMSC}(mdl, cores = 1, ...)
}
\arguments{
\item{mdl}{An object of class \code{BMSC}, resulting from the
\link{BMSC} function.}

\item{cores}{The number of cores fo the \link{loo::relative_eff} function}

\item{...}{further arguments passed to the `loo::extract_log_lik` function.}
}
\value{
a list with the log likelihood of the single case and the
        control group,
        the MCMC effective sample size divided by the total sample size,
        and the leave-one-out cross-validation.
}
\description{
bmscstan wrapper for computing approximate leave-one-out cross-validation
using PSIS-LOO
for the single case and the control group
}
