% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-plot.R
\name{pp_check.BMSC}
\alias{pp_check.BMSC}
\title{Posterior predictive checks for BMSC objects}
\usage{
pp_check.BMSC(object, type = "dens", limited = TRUE)
}
\arguments{
\item{object}{a BMSC object}

\item{type}{\describe{
    \item{dens} density overlay plot
    \item{hist} histogram plot
    \item{mean} the distribution of the mean statistic, over the simulated datasets, compared to the mean of the real data
}}

\item{limited}{logical. TRUE if the output should be limited within 4 standard deviations, FALSE it should not. default TRUE}
}
\value{
a ggplot2 object
}
\description{
\code{pp_check()} plots the posterior predictive check for BMSC objects.
}
