% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-pairwise.R
\name{print.pairwise.BMSC}
\alias{print.pairwise.BMSC}
\title{Print summaries of Pairwise Bayesian Multilevel Single Case objects}
\usage{
\method{print}{pairwise.BMSC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pairwise.BMSC}, resulting from the \link{pairwise.BMSC} function.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print summaries of Pairwise Bayesian Multilevel Single Case objects
}
