% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-generic.R
\name{print.summary.BMSC}
\alias{print.summary.BMSC}
\title{Print summaries of Bayesian Multilevel Single Case objects}
\usage{
\method{print}{summary.BMSC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.BMSC}, resulting from the \link{summary.BMSC} function.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print summaries of Bayesian Multilevel Single Case objects
}
