% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-generic.R
\name{summary.BMSC}
\alias{summary.BMSC}
\title{Summarizing Bayesian Multilevel Single Case objects}
\usage{
\method{summary}{BMSC}(object, ...)
}
\arguments{
\item{object}{An object of class \code{BMSC}, resulting from the \link{\code{BMSC}} function.}

\item{...}{other arguments are ignored.}
}
\description{
summary method for class "BMSC".
}
