% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSC-loo.R
\name{waic.BMSC}
\alias{waic.BMSC}
\title{bmscstan wrapper to compute the Widely applicable information criterion
(WAIC)
for the single case and the control group}
\usage{
\method{waic}{BMSC}(mdl, ...)
}
\arguments{
\item{mdl}{An object of class \code{BMSC}, resulting from the
\link{BMSC} function.}

\item{...}{further arguments passed to the `loo::extract_log_lik` function.}
}
\value{
a list with the log likelihood of the single case
        and the control group,
        and the waic scores.
}
\description{
bmscstan wrapper to compute the Widely applicable information criterion
(WAIC)
for the single case and the control group
}
